% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_disruption_features}
\alias{build_disruption_features}
\title{Build Disruption Features for Multiple ITS Models}
\usage{
build_disruption_features(data, time_name, start_indices, lengths = NULL)
}
\arguments{
\item{data}{A data frame or tibble containing the time series.}

\item{time_name}{A string giving the name of the time variable in `data`.}

\item{start_indices}{A numeric vector of length K stating the start time indices of each intervention.}

\item{lengths}{A numeric vector of length K stating the duration of each intervention. 
If NULL, disruptions are assumed to be permanent (lasting until the end of the series).}
}
\value{
The original data frame with 2*K new columns appended (indicator_k and shifted_time_k).
}
\description{
This helper function takes a dataframe and user inputs for multiple disruptions,
returning the dataframe appended with K level indicators and K shifted time continuous vectors.
}
