% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_lm}
\alias{its_lm}
\title{ITS analysis for continuous outcomes with multiple interruptions}
\usage{
its_lm(
  data,
  form,
  time_name,
  intervention_start_indices,
  intervention_lengths = NULL,
  multi_model_type = "singular",
  freq = NULL,
  seasonality = "none",
  keep_significant_fourier = TRUE,
  impact_model = "full",
  counterfactual = FALSE,
  print_summary = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the continuous outcome, and (2) a vector of time points.}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right. Should not contain an offset term.}

\item{time_name}{A string giving the name of the time variable. The time variable may or may not be supplied as a covariate in the formula.}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent. Default is NULL.}

\item{multi_model_type}{A string specifying the multi-disruption architecture. Options are "singular", "multiple", or "mixed". Default is "singular". Note: "mixed" is not supported if seasonality is set to "full".}

\item{freq}{A positive integer describing the frequency of the time series (e.g., 12 for monthly). Required if seasonality is "full". Default is NULL.}

\item{seasonality}{A string specifying whether seasonality should be considered. Options are "none" (no seasonal adjustment) or "full" (Fourier terms applied). Default value is "none".}

\item{keep_significant_fourier}{Logical - if seasonality is "full", indicates whether only significant Fourier terms should be retained. Default is TRUE.}

\item{impact_model}{A string specifying the assumed impact model: "full", "level", or "slope". Default value is "full".}

\item{counterfactual}{Logical - indicating whether the model-based counterfactual values should also be returned. Default is FALSE.}

\item{print_summary}{Logical - indicating whether the entire model summary should be printed. Default is FALSE.}
}
\value{
The function returns a list with three elements: the fitted linear regression model, the summary of the model (including Cohen's d), and the original data together with the model predictions.
}
\description{
\code{its_lm} fits a linear regression model to an Interrupted Time Series with one or more disruptions. 
It acts as a unified wrapper, automatically routing the data to either a standard linear model or a 
Fourier-adjusted seasonal model based on user input. It returns the fitted model, the summary 
(including Cohen's d), and the original data appended with model predictions.
}
\examples{
data <- unemployed
form <- as.formula("percent ~ time")
start_indices <- c(which(data$year==2020 & data$month==3), which(data$year==2021 & data$month==1))
fit <- its_lm(data=data, form=form, time_name="time", 
              intervention_start_indices=start_indices, freq=12, seasonality="none", 
              multi_model_type="multiple", impact_model="full", counterfactual=TRUE)
}
