% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_poisson}
\alias{its_poisson}
\title{ITS analysis for count outcomes with multiple interruptions}
\usage{
its_poisson(
  data,
  form,
  offset_name = NULL,
  time_name,
  intervention_start_indices,
  intervention_lengths = NULL,
  multi_model_type = "singular",
  freq = NULL,
  seasonality = "none",
  keep_significant_fourier = TRUE,
  over_dispersion = FALSE,
  impact_model = "full",
  counterfactual = FALSE,
  print_summary = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the count outcome, and (2) a vector of time points.}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right. Should not contain an offset term.}

\item{offset_name}{A string indicating the name of the offset column in the data, or NULL. Default value is NULL.}

\item{time_name}{A string giving the name of the time variable.}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent.}

\item{multi_model_type}{A string specifying the multi-disruption architecture: "singular", "multiple", or "mixed". Default is "singular".}

\item{freq}{A positive integer describing the frequency of the time series. Required if seasonality is "full".}

\item{seasonality}{A string specifying whether seasonality should be considered: "none" or "full". Default is "none".}

\item{keep_significant_fourier}{Logical - if seasonality is "full", indicates whether only significant Fourier terms should be retained. Default is TRUE.}

\item{over_dispersion}{Logical - if TRUE, uses quasi-Poisson. Default is FALSE. (Note: ignored for mixed-effects models).}

\item{impact_model}{A string specifying the impact model: "full", "level", or "slope". Default is "full".}

\item{counterfactual}{Logical - indicates whether model-based counterfactual values should be returned. Default is FALSE.}

\item{print_summary}{Logical - indicates whether the model summary should be printed. Default is FALSE.}
}
\value{
A list with three elements: the fitted model, the model summary (including relative risk), and the data with predictions.
}
\description{
\code{its_poisson} acts as a wrapper to fit either a standard or Fourier-adjusted 
Poisson (or quasi-Poisson) regression model to an ITS with one or more disruptions.
}
\examples{
data <- unemployed
form <- as.formula("unemployed ~ time")
start_indices <- c(which(data$year==2020 & data$month==3), which(data$year==2021 & data$month==1))
fit <- its_poisson(data=data, form=form, offset_name="labour", time_name="time", 
                   intervention_start_indices=start_indices, multi_model_type="multiple", 
                   seasonality="none", impact_model="full", counterfactual=TRUE)
}
