% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_poisson_wo_seas}
\alias{its_poisson_wo_seas}
\title{ITS analysis for count outcomes with multiple interruptions (no seasonality adjustment)}
\usage{
its_poisson_wo_seas(
  data,
  form,
  offset_name = NULL,
  time_name,
  intervention_start_indices,
  intervention_lengths = NULL,
  multi_model_type = "singular",
  over_dispersion = FALSE,
  impact_model = "full",
  counterfactual = FALSE,
  print_summary = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the count outcome, and (2) a vector of time points.}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right, separated by + operators. The formula should not contain an offset term.}

\item{offset_name}{A string indicating the name of the offset column in the data, or NULL. Default value is NULL.}

\item{time_name}{A string giving the name of the time variable. The time variable may or may not be supplied as a covariate in the formula.}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent (lasting until the end of the series). Default is NULL.}

\item{multi_model_type}{A string specifying the multi-disruption architecture. Options are "singular" (homogenous effects), "multiple" (distinct effects), or "mixed" (hierarchical effects). Default is "singular".}

\item{over_dispersion}{Logical - indicating whether a quasi-Poisson model should be used to account for overdispersion (TRUE), or not (FALSE). Default value is FALSE. Note: This argument is ignored if multi_model_type is "mixed" due to lme4 limitations.}

\item{impact_model}{A string specifying the assumed impact model. Options include "full" (level and slope change), "level" (just level change), and "slope" (just slope change). Default value is "full".}

\item{counterfactual}{Logical - indicating whether the model-based counterfactual values should also be returned as an additional column in the data. Default value is FALSE.}

\item{print_summary}{Logical - indicating whether the entire model summary should be printed, or just the relevant effect size. Default value is FALSE.}
}
\value{
The function returns a list with three elements: the fitted Poisson regression model, the summary of the model (including the relative risk), and the original data together with the model predictions.
}
\description{
\code{its_poisson_wo_seas} fits a Poisson regression model to an ITS with one or more disruption events. 
It returns the fitted model, the summary (including the relative risk), and the original data 
appended with the model's factual and counterfactual predictions.
}
\examples{
data <- unemployed
form <- as.formula("unemployed ~ time")
# Simulating two disruptions
start_indices <- c(which(data$year==2020 & data$month==3), which(data$year==2021 & data$month==1))
lengths <- c(6, 12) 
fit <- its_poisson_wo_seas(data=data, form=form, offset_name="labour", time_name="time", 
                           intervention_start_indices=start_indices, intervention_lengths=lengths, 
                           multi_model_type="multiple", over_dispersion=TRUE, 
                           impact_model="full", counterfactual=TRUE)
}
