% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{its_zero_inflated_fourier}
\alias{its_zero_inflated_fourier}
\title{ITS analysis for zero-inflated count outcomes with seasonal adjustment}
\usage{
its_zero_inflated_fourier(
  data,
  form,
  offset_name = NULL,
  time_name,
  intervention_start_indices,
  intervention_lengths = NULL,
  multi_model_type = "singular",
  freq,
  impact_model = "full",
  counterfactual = FALSE,
  print_summary = FALSE
)
}
\arguments{
\item{data}{The data frame corresponding to the supplied formula, existing of at least 2 variables: (1) the count outcome, and (2) a vector of time points.}

\item{form}{A formula with the response on the left, followed by the ~ operator, and the covariates on the right. Should not contain an offset term.}

\item{offset_name}{A string indicating the name of the offset column in the data, or NULL. Default value is NULL.}

\item{time_name}{A string giving the name of the time variable. The time variable may or may not be supplied as a covariate in the formula.}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent.}

\item{multi_model_type}{A string specifying the multi-disruption architecture. Options are "singular", "multiple", or "mixed". Default is "singular".}

\item{freq}{A positive integer describing the frequency of the time series (e.g., 12 for monthly data).}

\item{impact_model}{A string specifying the assumed impact model: "full", "level", or "slope". Default value is "full".}

\item{counterfactual}{Logical - indicating whether the model-based counterfactual values should also be returned. Default is FALSE.}

\item{print_summary}{Logical - indicating whether the entire model summary should be printed. Default is FALSE.}
}
\value{
The function returns a list with three elements: the fitted zero-inflated Poisson model, the summary of the model (including the relative risk), and the original data together with the model predictions.
}
\description{
\code{its_zero_inflated_fourier} fits a zero-inflated Poisson regression model adjusted for seasonality 
to an ITS with multiple interruptions. It returns the fitted model, the summary (including relative risk), 
and the original data appended with model predictions.
Note: Combining mixed-effects multi-disruption architectures with Fourier terms and zero-inflation is 
computationally heavy and may be prone to non-convergence.
}
\examples{
data <- zero_inflated_sim_data
form <- as.formula("monthly_total ~ time")
start_indices <- c(which(data$Year==2020 & data$Month==3), which(data$Year==2021 & data$Month==1))
fit <- its_zero_inflated_fourier(data=data, form=form, time_name="time", 
                                 intervention_start_indices=start_indices, freq=12, 
                                 multi_model_type="multiple", impact_model="full", counterfactual=TRUE)
}
