% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_its_lm}
\alias{plot_its_lm}
\title{Plot ITS fitted values with multiple interruptions}
\usage{
plot_its_lm(
  data,
  intervention_start_indices,
  intervention_lengths = NULL,
  y_lab,
  response,
  date_name
)
}
\arguments{
\item{data}{The data frame containing the original data and the model predictions (returned by the its_lm function).}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent.}

\item{y_lab}{A string with the y-axis label for the plot.}

\item{response}{A string giving the name of the response variable to be plotted in the scatterplot.}

\item{date_name}{A string giving the name of the date column. The date column must be a Date object.}
}
\value{
A ggplot object including a scatterplot of the time series, the predictions line, and the counterfactual predictions in red (if available).
}
\description{
\code{plot_its_lm} uses ggplot2 to plot the model-based fitted values, together with a scatterplot 
of the observed time series. It supports multiple disruptions, automatically drawing broken regression 
lines and vertical segments at the start and end of each interruption.
}
\examples{
\dontrun{
data <- unemployed
form <- as.formula("percent ~ time")
start_indices <- c(which(data$year==2020 & data$month==3), which(data$year==2021 & data$month==1))
fit <- its_lm(data=data, form=form, time_name="time", 
              intervention_start_indices=start_indices, 
              multi_model_type="multiple", counterfactual=TRUE)

p <- plot_its_lm(data=fit$data, intervention_start_indices=start_indices, 
                 y_lab="Unemployment percent", response="percent", date_name="dt")
print(p)
}
}
