% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_its_poisson}
\alias{plot_its_poisson}
\title{Plot ITS fitted values for count outcomes with multiple interruptions}
\usage{
plot_its_poisson(
  data,
  intervention_start_indices,
  intervention_lengths = NULL,
  y_lab,
  response,
  offset_name = NULL,
  date_name
)
}
\arguments{
\item{data}{The data frame containing the original data and the model predictions (returned by its_poisson or its_zero_inflated_poisson).}

\item{intervention_start_indices}{Numeric vector - stating the time point indices of the start of each intervention.}

\item{intervention_lengths}{Numeric vector - stating the duration of each intervention. If NULL, disruptions are assumed to be permanent.}

\item{y_lab}{A string with the y-axis label for the plot.}

\item{response}{A string giving the name of the response variable to be plotted in the scatterplot.}

\item{offset_name}{A string indicating the name of the offset column in the data, or NULL. If provided, values are normalized as (value * 100 / offset).}

\item{date_name}{A string giving the name of the date column. The date column must be a Date object.}
}
\value{
A ggplot object including a scatterplot of the time series, the predictions line, and the counterfactual predictions in red (if available).
}
\description{
\code{plot_its_poisson} uses ggplot2 to plot the model-based fitted values, together with a scatterplot 
of the observed time series. If an offset is provided, it automatically normalizes the predictions 
and observations into a standardized rate. It supports multiple disruptions by drawing broken 
regression lines and vertical segments at the boundaries of each interruption.
}
\examples{
\dontrun{
data <- unemployed 
form <- as.formula("unemployed ~ time")
start_indices <- c(which(data$year==2020 & data$month==3), which(data$year==2021 & data$month==1))
fit <- its_poisson(data=data, form=form, offset_name="labour", time_name="time", 
                   intervention_start_indices=start_indices, 
                   multi_model_type="multiple", counterfactual=TRUE)

p <- plot_its_poisson(data=fit$data, intervention_start_indices=start_indices, 
                      y_lab="Unemployment Rate", response="unemployed", 
                      offset_name="labour", date_name="dt")
print(p)
}
}
